+++ src/insight/insights.module.ts
@@ -6,6 +6,7 @@ import { DbInsightRepo } from "./entities/insight-repo.entity";
 import { InsightController } from "./insight.controller";
 import { UserInsightsController } from "./user-insight.controller";
 import { InsightsService } from "./insights.service";
+import { InsightRepoService } from "./insight-repo.service";
 
 @Module({
   controllers: [InsightController, UserInsightsController],
@@ -15,7 +16,7 @@ import { InsightsService } from "./insights.service";
       DbInsightRepo,
     ]),
   ],
-  providers: [InsightsService],
-  exports: [InsightsService],
+  providers: [InsightsService, InsightRepoService],
+  exports: [InsightsService, InsightRepoService],
 })
 export class InsightsModule {}